Remove mention of --no-explicit-require from readme
This mentions "These three scripts" when there are four mentioned, and
I am unable to find any references to `--no-explicit-require` in code,
so I think this sentence is outdated and should be removed. Digging
through history, it seems like this was brought over from 1c672cbb9f9
when code was imported from the React repo.